+++ playlistparser/__init__.py
@@ -296,4 +296,4 @@ __all__ = [
     "UnknownFormatError",
 ]
 
-__version__ = "4.1.0"
+__version__ = "4.3.0"